Swap names of charts and index
and renamed the new charts to plan projects
