--- AcademicTimePlanner/Pages/Charts.razor
@@ -1,31 +1,0 @@
-@page "/charts"
-@inherits Fluxor.Blazor.Web.Components.FluxorComponent
-
-@using Plotly.Blazor.LayoutLib
-
-@if (ChartsState.Value.Loaded)
-{
-    <PlotlyChart Id="Total" Config="config" Layout="layoutTotal" Data=GetDataTotal()/>
-    <PlotlyChart Id="Projects" Config="config" Layout="layoutProjects" Data=GetDataOfSingleProjectsToday()/>
-
-    @if (!ChartsState.Value.LoadedFiltered)
-    {
-        <EditForm Model="@DateFilter" OnValidSubmit="@SetDateFilter">
-            <DataAnnotationsValidator />
-            <ValidationSummary />
-
-            <label>Start date: <InputDate id="StartDate" @bind-Value="DateFilter.StartDate" /></label>
-            <label>End date: <InputDate id="EndDate" @bind-Value="DateFilter.EndDate" /></label>
-            <button type="submit">Set filter</button>
-        </EditForm>
-    }
-    else
-    {
-        @foreach(var project in ChartData.PlanProjects)
-        {
-            <PlotlyChart Id="@project.Name" Config="config" Layout="GetLayout(project.Name)" Data="GetDataOfSingleProjectsFiltered(project)" />
-        }        
-        <button type="button" class="btn btn-secondary" @onclick="ChangeFilter">Reset filter</button>
-    }
-}
-
