+++ AcademicTimePlanner/Pages/Index.razor.cs
@@ -1,116 +1,251 @@
 using AcademicTimePlanner.Data;
 using AcademicTimePlanner.DataMapping.Plan;
-using AcademicTimePlanner.Store.State.ProjectFiles;
+using AcademicTimePlanner.Store.State.Charts;
 using AcademicTimePlanner.Store.State.Wrapper;
 using Fluxor;
 using Microsoft.AspNetCore.Components;
-using Microsoft.AspNetCore.Components.Forms;
-using Microsoft.JSInterop;
-using Newtonsoft.Json;
+using Plotly.Blazor;
+using Plotly.Blazor.LayoutLib;
+using Plotly.Blazor.Traces.ScatterLib;
+using Bar = Plotly.Blazor.Traces.Bar;
+using BarMarker = Plotly.Blazor.Traces.BarLib.Marker;
+using LineMarker = Plotly.Blazor.Traces.ScatterLib.Marker;
+using Scatter = Plotly.Blazor.Traces.Scatter;
+using YAxisTitle = Plotly.Blazor.LayoutLib.YAxisLib.Title;
 
 namespace AcademicTimePlanner.Pages;
 
 public partial class Index
 {
     [Inject]
-    private IState<ProjectFilesState> ProjectState { get; set; }
+    private IState<ChartsState> ChartsState { get; set; }
 
     [Inject]
     private IDispatcher Dispatcher { get; set; }
 
-    [Inject]
-    public IJSRuntime JSRuntime { get; set; }
+    private ChartData? ChartData => ChartsState.Value.ChartData;
+    private DateFilter? DateFilter => ChartsState.Value.DateFilter;
 
-    private List<string> PlanProjectsNames => ProjectState.Value.PlanProjectsNames;
+    private const string Title = "Charts";
+    private const string TotalChartTitle = "Total";
 
-    private const string Title = "Plan projects";
+    private static readonly BarMarker TrackedDurationMarker = new BarMarker { Color = "rgb(20, 150, 70)" };
+    private static readonly BarMarker PlannedDurationMarker = new BarMarker { Color = "rgb(20, 70, 150)" };
+    private static readonly BarMarker PredictedDurationMarker = new BarMarker { Color = "rgb(34, 220, 93)" };
+    private static readonly BarMarker TotalDurationMarker = new BarMarker { Color = "rgb(34, 120, 250)" };
 
-    private PlanProject? planProject => ProjectState.Value.PlanProject;
+    private static readonly LineMarker TrackedDurationLine = new LineMarker { Color = "rgb(20, 150, 70)" };
+    private static readonly LineMarker PlannedDurationLine = new LineMarker { Color = "rgb(20, 70, 150)" };
 
-    private PlanTask? planTask => ProjectState.Value.PlanTask;
+    Config config = new Config
+    {
+        Responsive = true
+    };
 
-    private PlanEntry? planEntry => ProjectState.Value.PlanEntry;
+    Layout layoutTotal = new Layout
+    {
+        Title = new Title
+        {
+            Text = "Total overview"
+        },
+        BarMode = BarModeEnum.Group,
+        XAxis = new List<XAxis> { new XAxis { Anchor = "free", Position = 0 }, new XAxis { Anchor = "free", Position = 0, Overlaying = "x" } },
+        YAxis = new List<YAxis> { new YAxis { Title = new YAxisTitle { Text = "hours" } } },
+        Height = 500,
+        Width = 700,
+    };
 
-    private PlanEntryRepetition? planEntryRepetition => ProjectState.Value.PlanEntryRepetition;
+    Layout layoutProjects = new Layout
+    {
+        Title = new Title
+        {
+            Text = "Projects overview"
+        },
+        BarMode = BarModeEnum.Group,
+        XAxis = new List<XAxis> { new XAxis { Anchor = "free", Position = 0, TickAngle = 45 }, new XAxis { Anchor = "free", Position = 0, Overlaying = "x", TickAngle = 45 } },
+        YAxis = new List<YAxis> { new YAxis { Title = new YAxisTitle { Text = "hours" } } },
+        Height = 500,
+        AutoSize = true,
+        BarGroupGap = 0
+    };
 
-    private PlanProjectDownloader downloader => ProjectState.Value.PlanProjectDownloader;
+    private Layout GetLayout(string projectName)
+    {
+        return new Layout
+        {
+            Title = new Title
+            {
+                Text = projectName + " overview in time range (" + DateFilter.StartDate.Date + " - " + DateFilter.EndDate.Date + ")"
+            },
+            YAxis = new List<YAxis> { new YAxis { Title = new YAxisTitle { Text = "hours" } } },
+            Height = 500,
+            AutoSize = true,
+        };
+    }
+
+    private List<ITrace> GetDataTotal()
+    {
+        return new List<ITrace>
+        {
+            new Bar
+            {
+                X = new List<object> {TotalChartTitle},
+                Y = new List<object> {ChartData!.TotalTrackedTime + ChartData!.RemainingDuration},
+                Name = "Total predicted",
+                Marker = PredictedDurationMarker,
+            },
+            new Bar
+            {
+                X = new List<object> {TotalChartTitle},
+                Y = new List<object> {ChartData!.TotalPlannedTime},
+                Name = "Total planned",
+                XAxis = "x2",
+                Marker = PlannedDurationMarker,
+
+            },
+            new Bar
+            {
+                X = new List<object> {TotalChartTitle},
+                Y = new List<object> {ChartData!.TotalTrackedTime},
+                Name = "Total tracked",
+                XAxis = "x2",
+                Marker = TrackedDurationMarker,
+            },
+        };
+    }
+
+    private List<ITrace> GetDataOfSingleProjectsToday()
+    {
+        var titles = new List<object>();
+        var totalDurations = new List<object>();
+        var predictedDurations = new List<object>();
+        var plannedDurations = new List<object>();
+        var trackedDurations = new List<object>();
+
+        foreach (var planProject in ChartData!.PlanProjects)
+        {
+            double predictedDurationsSum = 0;
+            double trackedDurationsSum = 0;
+            foreach (long togglProjectId in planProject.TogglProjectIds.Keys)
+            {
+                var togglProject = ChartData!.GetTogglProjectWithTogglId(togglProjectId);
+                if (togglProject != null)
+                {
+                    predictedDurationsSum += togglProject.GetTotalDuration() * planProject.TogglProjectIds[togglProjectId];
+                    trackedDurationsSum += togglProject.GetTotalDuration() * planProject.TogglProjectIds[togglProjectId];
+                }
+                predictedDurationsSum += planProject.GetRemainingDuration();
+            }
+            titles.Add(planProject.Name);
+
+            totalDurations.Add(planProject.GetTotalDuration());
+            predictedDurations.Add(predictedDurationsSum);
+            plannedDurations.Add(planProject.GetTotalDuration() - planProject.GetRemainingDuration());
+            trackedDurations.Add(trackedDurationsSum);
+        }
+
+        return new List<ITrace>
+        {
+            new Bar
+            {
+                X = titles,
+                Y = totalDurations,
+                Name = "Predicted",
+                Marker = TotalDurationMarker,
+            },
+            new Bar
+            {
+                X = titles,
+                Y = predictedDurations,
+                Name = "Predicted",
+                Marker = PredictedDurationMarker,
+            },
+            new Bar
+            {
+                X = titles,
+                Y = plannedDurations,
+                Name = "Planned",
+                XAxis = "x2",
+                Marker = PlannedDurationMarker,
+            },
+            new Bar
+            {
+                X = titles,
+                Y = trackedDurations,
+                Name = "Tracked",
+                XAxis = "x2",
+                Marker = TrackedDurationMarker,
+            },
+        };
+    }
+
+    private List<ITrace> GetDataOfSingleProjectsFiltered(PlanProject planProject)
+    {
+        var plannedDurations = planProject.GetDurationsPerDateInTimeRange(DateFilter.StartDate, DateFilter.EndDate);
+
+        var plannedDurationsDates = new List<object>();
+        var plannedDurationsTimes = new List<object>();
+        plannedDurations.Keys.ToList().ForEach(date => plannedDurationsDates.Add(date));
+        plannedDurations.Values.ToList().ForEach(time => plannedDurationsTimes.Add(time));
+
+        var trackedDurationsDates = new List<object>();
+        var trackedDurationsTimes = new List<object>();
+        SortedDictionary<DateTime, double> trackedDurations = new SortedDictionary<DateTime, double>();
+        List<long> togglProjects = new List<long>();
+        togglProjects.AddRange(planProject.TogglProjectIds.Keys.ToList());
+
+        for (int i = 0; i < togglProjects.Count; i++)
+        {
+            var togglProject = ChartData!.GetTogglProjectWithTogglId(togglProjects[i]);
+            var durationsPerDate = togglProject.GetDurationsPerDate(trackedDurations, planProject.TogglProjectIds[togglProjects[i]]);
+            if (i != planProject.TogglProjectIds.Count - 1)
+            {
+                trackedDurations = togglProject.GetDurationsPerDateInTimeRange(DateTime.MinValue, DateTime.MaxValue, durationsPerDate);
+            }
+            else
+            {
+                trackedDurations = togglProject.GetDurationsPerDateInTimeRange(DateFilter.StartDate, DateFilter.EndDate, togglProject.Sumup(durationsPerDate));
+            }
+        }
+        trackedDurations.Keys.ToList().ForEach(date => trackedDurationsDates.Add(date));
+        trackedDurations.Values.ToList().ForEach(time => trackedDurationsTimes.Add(time));
+
+        return new List<ITrace>
+        {
+            new Scatter
+            {
+                X = plannedDurationsDates,
+                Y = plannedDurationsTimes,
+                Name = "Planned",
+                Mode = ModeFlag.Lines| ModeFlag.Markers,
+                Marker = PlannedDurationLine,
+            },
+            new Scatter
+            {
+                X = trackedDurationsDates,
+                Y = trackedDurationsTimes,
+                Name = "Tracked",
+                Mode = ModeFlag.Lines| ModeFlag.Markers,
+                Marker = TrackedDurationLine,
+            },
+        };
+    }
 
     protected override void OnInitialized()
     {
         base.OnInitialized();
+        ChartsState.Value.Loaded = false;
         Dispatcher.Dispatch(new SetTitleAction(Title));
+        Dispatcher.Dispatch(new FetchChartDataAction());
     }
 
-    private async Task LoadPlanProjects(InputFileChangeEventArgs e)
+    private void SetDateFilter()
     {
-        var json = new List<string>();
-        foreach (var file in e.GetMultipleFiles(int.MaxValue))
-        {
-            json.Add(await new StreamReader(file.OpenReadStream()).ReadToEndAsync());
-        }
-        Dispatcher.Dispatch(new LoadPlanProjectsAction(json));
+        Dispatcher.Dispatch(new FilterChartDataAction());
     }
 
-    private void CreatePlanProject()
+    private void ChangeFilter()
     {
-        Dispatcher.Dispatch(new SwitchCreationStepAction(ProjectFilesState.CreationStep.NamingProject, new PlanProject(Guid.NewGuid())));
-    }
-
-    private void CreatePlanTask()
-    {
-        Dispatcher.Dispatch(new CreatePlanTaskAction(planTask));
-    }
-
-    private void CreatePlanEntry()
-    {
-        planProject.AddPlanEntry(planEntry);
-        Dispatcher.Dispatch(new AddSingleEntryAction());
-    }
-
-    private void CreateRepetitionEntry()
-    {
-        planProject.AddRepetitionEntry(planEntryRepetition);
-        Dispatcher.Dispatch(new AddRepetitionEntryAction());
-    }
-
-    private void AddSingleEntry()
-    {
-        Dispatcher.Dispatch(new AddSingleEntryAction());
-    }
-
-    private void AddRepetitionEntry()
-    {
-        Dispatcher.Dispatch(new AddRepetitionEntryAction());
-    }
-
-    private void Cancel()
-    {
-        Dispatcher.Dispatch(new SwitchCreationStepAction(ProjectFilesState.CreationStep.NotCreating, null));
-    }
-
-    private void NextOrBack(ProjectFilesState.CreationStep step)
-    {
-        Dispatcher.Dispatch(new SwitchCreationStepAction(step, planProject!));
-    }
-
-    private void Finish()
-    {
-        Dispatcher.Dispatch(new FinishPlanProjectCreationAction(planProject!));
-    }
-
-    private void InitializePlanProjectDownload()
-    {
-        Dispatcher.Dispatch(new GetPlanProjectForDownloadAction(downloader.ProjectName));
-    }
-
-    private async Task DownloadPlanProject()
-    {
-        var fileStream = new MemoryStream();
-        var writer = new StreamWriter(fileStream);
-        writer.Write(JsonConvert.SerializeObject(planProject, Formatting.Indented));
-        writer.Flush();
-        fileStream.Position = 0;
-        using var streamRef = new DotNetStreamReference(stream: fileStream);
-        await JSRuntime.InvokeVoidAsync("downloadFileFromStream", planProject!.Name + ".json", streamRef);
+        Dispatcher.Dispatch(new ChangeFilterAction());
     }
 }
