+++ AcademicTimePlanner/Shared/NavMenu.razor
@@ -11,12 +11,12 @@
     <nav class="flex-column">
         <div class="nav-item px-3">
             <NavLink class="nav-link" href="" Match="NavLinkMatch.All">
-                <span class="oi oi-home" aria-hidden="true"></span>Index
+                <span class="oi oi-home" aria-hidden="true"></span>Charts
             </NavLink>
         </div>
        <div class="nav-item px-3">
-           <NavLink class="nav-link" href="charts">
-               <span class="oi oi-list-rich" aria-hidden="true"></span>Charts
+           <NavLink class="nav-link" href="planProjects">
+               <span class="oi oi-list-rich" aria-hidden="true"></span>Plan projects
            </NavLink>
        </div>
         <div class="nav-item px-3">
